chore: query refactoring
