+++ app/components/ui/FavoritePill.tsx
@@ -1,13 +1,13 @@
-import React, { memo } from 'react';
-import { TouchableOpacity } from 'react-native';
 import { FontAwesome } from '@expo/vector-icons';
-import IconPill from './IconPill';
-import useAppStore from '../../data/app_state';
 import {
   addFavoriteMutation,
   removeFavoriteMutation,
   useFavorite,
-} from 'app/data/storage_query';
+} from 'app/data/queries/structure/storage';
+import React, { memo } from 'react';
+import { TouchableOpacity } from 'react-native';
+import useAppStore from '../../data/app_state';
+import IconPill from './IconPill';
 
 interface Props {
   routeShortName: string;
